% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/github-meta.R
\name{rdcm}
\alias{rdcm}
\alias{use_rdcm_contributing}
\alias{use_rdcm_coc}
\title{Helpers for r-dcm development}
\usage{
use_rdcm_contributing()

use_rdcm_coc()
}
\description{
These helpers are adapted from \link[usethis:tidyverse]{usethis::tidyverse} to apply common
conventions to r-dcm packages.
}
\details{
\itemize{
\item \code{use_rdcm_contributing()}: adds standard r-dcm contributing guidelines.
\item \code{use_rdcm_coc()}: equivalent to \code{\link[usethis:use_code_of_conduct]{usethis::use_code_of_conduct()}}, but puts
the document in a \verb{.github/} subdirectory.
}
}
