% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/userdcm-package.R
\docType{package}
\name{userdcm-package}
\alias{userdcm}
\alias{userdcm-package}
\title{userdcm: Automate Package Setup}
\description{
Automate package setup tasks for r-dcm packages that are otherwise performed manually.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/r-dcm/userdcm}
  \item Report bugs at \url{https://github.com/r-dcm/userdcm/issues}
}

}
\author{
\strong{Maintainer}: W. Jake Thompson \email{wjakethompson@gmail.com} (\href{https://orcid.org/0000-0001-7339-0300}{ORCID})

Other contributors:
\itemize{
  \item University of Kansas [copyright holder]
  \item Institute of Education Sciences [funder]
}

}
\keyword{internal}
